refactor: register target flags per cmd instead of globally
